Allow Editors to use iFrame and Script tags
Generally this is a no-no, but the reason this wasn't working for them
is strictly because they are not Admins.

Since the Editors could come back in the future and edit this page, they
need to have this ability otherwise they'd end up wiping this change out
once they publish their changes.

Closes #112